move beamlines to module
