+++ tests/beamlines/unit_tests/test_i03.py
@@ -2,7 +2,7 @@ import pytest
 from ophyd import Device
 from ophyd.sim import FakeEpicsSignal
 
-from dodal import beamline_utils, i03
+from dodal.beamlines import beamline_utils, i03
 from dodal.devices.aperturescatterguard import AperturePositions, ApertureScatterguard
 from dodal.devices.smargon import Smargon
 from dodal.devices.zebra import Zebra
